docker/requirements.txt
uvicorn
fastapi
diffusers
matplotlib
loguru
accelerate
Pillow
pybase64
optimum[openvino]
